gloas: address §6 envelope/proposer review findings
- BlindedExecutionPayloadEnvelope.BuilderIndex: use the named gloas.BuilderIndex
  (as the bid does; SSZ-identical). The per-file sszgen directive can't resolve a
  sibling-file type, so switch it to package mode (-path .) like beacon_block.go.
- Note both ExecutionRequests fields track the pinned pre-EIP-8282 Gloas spec; a
  node-side variant follows when the target devnet adopts EIP-8282 (builder
  deposit/exit requests).
- submitGloasProposal: the self-build builder now triggers the §6 envelope even
  when its own block submission fails — the envelope is a cluster round the other
  operators join regardless, so the builder shouldn't abstain.
- startEnvelopeDuty: drop the post-consensus context it forwarded (cancelled when
  the block duty ends); document that the wired starter must dispatch async with a
  node-scoped context.

+++ protocol/v2/ssv/value_check.go
@@ -198,7 +198,7 @@ func (v *envelopeChecker) CheckValue(value []byte) error {
 	}
 
 	// This duty applies only to the self-build path; external builders sign their own envelopes.
-	if blinded.BuilderIndex != uint64(gloas.BuilderIndexSelfBuild) {
+	if blinded.BuilderIndex != gloas.BuilderIndexSelfBuild {
 		return spectypes.NewError(spectypes.QBFTValueInvalidErrorCode, "envelope builder index is not self-build")
 	}
